requirements.txt
streamlit>=1.28.0
pandas>=2.0.0
numpy>=1.24.0
plotly>=5.15.0
matplotlib>=3.7.0
pytest>=7.4.0
black>=23.10.0
flake8>=6.1.0
pyyaml>=6.0.0
tqdm>=4.66.0
scikit-learn>=1.3.0
transformers>=4.30.0
datasets>=2.14.0
beaker-py<2.0.0